misc: remove redundant type casting (#289)
There are several places with redundant type casting.
Presumably, many if not all of these were probably due
to constraints of the early Go compiler. Although there
is a certain readability aspect, there is potential over-
head in redundant casts. The Go compiler might add in
overhead to cast.

+++ hrpc/call.go
@@ -214,7 +214,7 @@ func cellFromCellBlock(b []byte) (*pb.Cell, uint32, error) {
 	key := b[:keyLen]
 	b = b[keyLen:]
 
-	familyLen := uint8(b[0])
+	familyLen := b[0]
 	b = b[1:]
 
 	family := b[:familyLen]
@@ -233,7 +233,7 @@ func cellFromCellBlock(b []byte) (*pb.Cell, uint32, error) {
 	timestamp := binary.BigEndian.Uint64(b[:8])
 	b = b[8:]
 
-	cellType := uint8(b[0])
+	cellType := b[0]
 	b = b[1:]
 
 	value := b[:valueLen]
